requirements.txt
--prefer-binary

# Web framework
flask==2.0.1
flask-cors==3.0.10
requests>=2.27.0

# Google API services
google-auth-oauthlib==0.4.6
google-api-python-client==2.0.2
google-analytics-data==0.16.0

# Social media integration
tweepy==4.10.0

# Image processing - using pre-built wheel
pillow==10.2.0 --only-binary :all:

# NLP and machine learning - using pre-built wheel
transformers==4.38.0 --only-binary :all:
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchvision==0.17.0+cpu
torchaudio==2.2.0+cpu

# Scheduling and utilities
schedule==1.1.0
python-dotenv==0.19.2
newsapi-python==0.2.6

fastapi
uvicorn[standard]
